Move make_numerical to tools
This is a generic function and should go in a generic file
