+++ inflation/sdp/quantum_tools.py
@@ -9,7 +9,7 @@ import sympy
 
 from itertools import permutations, product, combinations_with_replacement
 from tqdm import tqdm
-from typing import Any, Dict, List, Tuple, Union
+from typing import Dict, List, Tuple, Union
 
 from .fast_npa import (apply_source_perm,
                        dot_mon,
@@ -455,31 +455,3 @@ def lexicographic_order(infSDP) -> Dict[str, int]:
         lexorder[sympy.Symbol(to_name([op], infSDP.names),
                               commutative=False)] = i
     return lexorder
-
-###############################################################################
-# OTHER FUNCTIONS                                                             #
-###############################################################################
-def make_numerical(symbolic_expressions: Dict[Any, sympy.core.expr.Expr],
-                   symbols_to_values: Dict[sympy.core.symbol.Symbol, float]
-                   ) -> Dict[Any, float]:
-    """Replace the symbols in the values of a dictionary by the corresponding
-    numerical values.
-    Parameters
-    ----------
-    symbolic_expressions : Dict[Any, sympy.core.expr.Expr]
-        Dictionary where the values are symbolic expressions of some variables.
-    symbols_to_values : Dict[sympy.core.symbol.Symbol, float]
-        Correspondence of the variables in the expressions and their associated
-        numerical values.
-    Returns
-    -------
-    Dict[Any, float]
-        The dictionary with samy keys and evaluated expressions as values.
-    """
-    numeric_values = dict()
-    for k, v in symbolic_expressions.items():
-        try:
-            numeric_values[k] = float(v.evalf(subs=symbols_to_values))
-        except AttributeError:
-            numeric_values[k] = float(v)
-    return numeric_values
